edited.

+++ src/main/java/com/petclinic/repository/OwnerRepository.java
@@ -3,5 +3,5 @@ package com.petclinic.repository;
 import com.petclinic.model.Owner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OwnerRepository  extends JpaRepository<Owner,Long> {
+public interface OwnerRepository  extends JpaRepository<Owner,Integer> {
 }
